+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  - Added `luakit.install_paths` table. `luakit.install_path` is now deprecated.
  - Added `Control-Y` readline binding.
  - Added ability to control whether links from secondary instances open in a new window.
+ - Added `luakit.resource_path` property to control where luakit searches for resource files.
 
 ### Changed